PHOENIX-2110 Primary key changes should be pushed to diverged views
